+++ resources/data.go
@@ -69,7 +69,7 @@ func (fi bindataFileInfo) Sys() interface{} {
 	return nil
 }
 
-var _goCentrifugeBuildConfigsDefault_configYaml = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xc4\x58\x59\x73\xdc\x36\x12\x7e\x9f\x5f\xd1\x25\xbf\x24\x55\xcb\x11\x78\x93\x53\x95\xda\x9a\xd3\x76\x2c\x2b\xa3\xc3\x51\xac\x97\x35\x08\x36\x67\x60\x71\x00\x1a\x00\xe7\xf0\xaf\xdf\x02\xc8\x51\x24\xcb\x52\x76\x93\xca\xae\x5e\xc4\x01\xd0\x8d\x3e\xbe\xfe\xd0\xc0\x2b\x98\x61\x45\xdb\xda\x40\x89\x5b\xac\x65\xb3\x41\x61\xc0\xa0\x36\x02\x0d\xd0\x15\xe5\x42\x1b\x50\x5c\xdc\x61\x71\x18\x30\x14\x46\xf1\xaa\x5d\xe1\x39\x9a\x9d\x54\x77\x23\x50\xad\xd6\x9c\x8a\x35\xaf\xeb\x81\x53\xc6\x05\x82\x59\x23\x94\xbd\x5e\xd1\xad\xd4\x60\xd6\xd4\xc0\xf4\x5e\x03\x6c\x28\x17\xc6\xea\x1f\x1c\x97\x8c\x06\x00\xaf\xe0\x4c\x32\x5a\x3b\x13\xb8\x58\x01\x93\xc2\x28\xca\x0c\xd0\xb2\x54\xa8\x35\x6a\x10\x88\x25\x18\x09\x05\x82\x46\x03\x3b\x6e\xd6\x80\x62\x0b\x5b\xaa\x38\x2d\x6a\xd4\xc3\x01\x1c\xe5\xad\x4a\x00\x5e\x8e\x20\x0c\x43\xf7\x8d\x66\x8d\x0a\xdb\x4d\xef\xc1\xdb\x72\x04\x59\x98\x75\x73\x85\x94\x46\x1b\x45\x9b\x25\xa2\xd2\x9d\xac\x07\x27\xa7\xbc\x89\x4e\xfd\x20\x1d\x92\x21\x19\xfa\xa7\x86\x35\xa7\x61\x16\x90\xe0\x94\x37\x95\x3e\xbd\xd8\x5c\x5f\xec\x8b\xdd\x5d\x7b\xfb\xf1\xe3\xac\x6a\xbf\x5e\x17\xfb\xf9\xf8\x12\xaf\xcf\xa7\x67\xf2\xeb\xe1\x10\xc7\xd9\xf6\x42\xac\x7e\xdd\x2e\xdf\x7f\x3e\xfb\x78\x77\xf2\x07\x4a\xc3\xa3\xd2\x5f\xab\x64\x7e\x9e\x6c\xee\xbe\xdc\xe0\xe7\x9b\x77\x37\xc1\x97\x65\xeb\x27\xbf\x35\xe5\xeb\xf0\xee\x67\xe9\x5f\x87\x9b\x35\x5d\x2f\x27\xf1\x15\xc6\xc2\xef\x94\x1e\x43\x35\x3e\x46\xaa\x73\xc0\xba\x8f\xc2\x70\x73\x58\x50\x66\xa4\x3a\x8c\xe0\xe4\xe4\x9b\x99\x4b\x5c\x71\x6d\x1e\x4d\x51\xc1\xd6\x52\x5d\x62\x23\x35\xff\x46\xaa\xa1\x07\x0b\x93\x5f\x8a\x9a\xaf\xa8\xe1\x52\xb8\x39\x97\xbc\xf7\x94\x8b\xef\x42\xa9\xcf\x31\xfc\x70\xd9\x61\xe9\xc7\x01\x3c\xc4\x4e\x67\xea\x2b\x38\x6f\x37\xa8\x38\x83\xb7\x33\x90\x95\xc3\xd1\x03\xc4\xf4\x3a\xee\x53\x1a\xfb\xbd\xd4\xe4\x98\x37\xa8\xb9\x36\x56\x52\xc8\x12\x9f\x42\xae\x51\x72\xcb\xdd\x84\x74\xba\x1f\x18\x70\x34\xf4\x0f\x71\x10\xc6\xc3\x20\x88\x87\x01\x21\xc3\x28\xf8\x16\x0b\x7e\x30\x0b\xdf\x49\x79\x73\xc6\x39\xbb\xf8\x75\x77\xbd\xbe\x9e\x7c\x4c\xf6\xef\xd8\x52\x9e\x55\xc9\xe5\xc5\xc7\x9f\x17\xcd\xae\xf2\x55\x1a\xef\xce\xf6\xc1\xed\x65\xd8\x4c\x4b\xff\xe4\x7b\xea\xb3\x64\x18\xf8\xe4\x39\xf5\x17\xb7\xef\xc7\xd9\xeb\xe5\x1b\xb5\x9d\xdf\x4e\xf2\x5d\x79\x27\x3f\xb0\xf1\x78\x33\xbd\x7d\xd3\xe4\x78\x38\xdc\x46\x57\xf3\x6c\xb5\x50\xe1\xfa\xfa\xfc\xb7\x93\x3e\x46\xf3\x1e\xf7\xf7\x99\x78\x3b\x03\x0f\xfa\x6c\x3c\x57\x19\x51\x2f\x7c\x46\x6d\x78\xa0\xc4\xa6\x96\x07\x2c\xe1\x6a\x43\x95\x81\x69\x0f\x38\x0d\x95\x54\x2e\xa0\x2b\xbe\x45\xf1\x28\x94\x4f\x41\x09\xcf\xa2\x92\xec\x17\x61\x99\x67\x55\x58\x25\x59\x92\xcc\xd3\x34\x08\x49\x1c\x93\x28\xf4\xd3\xd8\xcf\xc7\x7e\x30\x63\x55\x86\x19\xe6\xe3\x17\xf0\x4b\xf6\x79\x92\x10\x46\xc2\xbc\x0c\x7d\x3f\x8a\x43\x5a\x91\x32\xce\x58\x9c\x24\x49\x1a\x84\x65\xce\x82\x8a\xa6\x65\x82\xec\x05\xa4\x93\x7d\x1c\xa6\xc1\x04\x93\x24\xa4\x65\xb8\x98\x21\x99\x04\x8b\x24\x4a\xd3\x84\x64\xfe\x94\xa4\xe9\x34\x9a\xc4\x94\x46\x71\xfa\x52\x4d\x90\x7d\xbc\x20\x41\x9e\x67\xf9\x6c\x9e\x8e\xc7\x11\x49\xe8\x8c\xf8\xb3\x99\x5f\x06\x71\x99\x54\x01\x92\xc0\x9f\x23\x12\x92\xf5\xd5\xf3\x4e\x6e\x69\x17\xbe\x07\x58\x2f\x50\x09\x5a\xaf\x91\xaf\xd6\xa6\xc7\xe2\xab\x57\xaf\xfa\xc4\x74\x12\x8b\xf1\x45\xff\xdb\x83\x1b\x4b\x87\x5c\x54\xad\xa2\x70\x90\x2d\xac\x2c\x8f\x0b\x40\xa5\xa4\xb2\x28\xbb\x5e\x73\x0d\x0a\xbf\xb4\x76\x17\xae\x41\x48\x03\xba\x6d\x1a\xa9\x0c\x96\x50\x20\xa3\xad\x46\x2b\xa9\x5c\x11\xd9\x25\xaa\x15\xc2\x72\xb1\x63\x5a\x6d\xa8\xb1\x95\xd4\xda\xa1\x21\x5c\xb6\xa2\x1b\xf7\xbc\x7e\xec\x27\xaa\xd8\x9a\x6f\x71\x78\xf2\x8f\xde\x28\x80\x9d\x2d\x44\x23\xa1\x94\xff\x74\x12\x14\x6a\xc7\xf2\x0d\x55\xdc\x1c\xba\x8d\x9c\x96\x3b\xe7\x0f\xae\x46\xdd\xcf\x4f\xfd\x02\xcf\x63\x6b\xca\xc5\x4f\xdd\xb4\xe7\x59\x6b\x7f\x0a\x49\x48\x22\xf0\xbc\x1d\x55\x4d\xff\xcf\x2b\xa8\x52\x1c\x15\xc4\x49\x46\x08\x21\xe0\x79\x42\x7a\x54\x30\x8e\xc2\x78\x45\x2d\xd9\x9d\xee\xc6\x34\xaa\x2d\x7a\xb5\x0d\x2a\x78\xde\x86\xee\xbd\xc6\xd6\x3a\x04\xb1\x15\xd2\x82\x36\x7a\x2d\x4d\x3f\xe8\xc6\x36\x5c\x3c\xfa\x69\x6d\xa6\xcc\xf0\x2d\x82\xe7\x59\x8c\xdb\x10\xc9\xaa\x7a\x1a\x09\xf0\xbc\xb2\xf0\x98\xdc\x34\x76\xbd\x14\xa0\x75\x69\x5d\xa2\x6c\x8d\x9e\xe6\x5f\x11\x22\x92\x27\xe0\x79\x9f\xb5\x14\xaa\x61\xde\x5a\x6a\xa3\x81\xd6\xf5\x83\x31\x2e\x0c\xaa\x8a\x32\xb4\xe3\x9f\x1e\xa7\xfb\x69\x30\xbf\x97\xf9\x89\x75\x1f\x4b\x5b\x92\x02\x3b\x43\x8c\x84\x1b\x2c\xae\xec\xb8\xd1\xe0\x62\xa2\xa0\x52\x72\x03\xad\x30\xaa\xd5\x16\x12\x52\xf1\x15\x17\x23\x18\x0e\x4f\x9e\xcd\xa7\xad\xfd\x27\xb9\xfc\xe4\x79\xad\xd0\xb4\x42\x0f\xf7\x8d\xd4\xf8\x09\xaa\x9a\xae\xbe\x01\xf0\x7f\x47\xf8\xc1\x5f\x24\xfc\x47\xb5\xf4\x1f\x53\xbe\x4f\xa2\xa1\x1f\x47\x43\x3f\x1b\xc6\x4f\x8e\xff\x23\x27\x2f\x75\xc2\x29\x7e\x68\x17\xb7\xe7\xad\xff\x7a\xbf\xd5\x87\xc9\xf5\x95\xba\xd6\xf9\xd6\x4c\x92\xc2\xbc\x1f\x8b\x37\x0b\x79\xf6\xb9\xb8\xfb\x3a\xa5\x27\xdf\x51\x1f\x0f\xfd\x2c\x1e\x06\x61\xfa\xec\x06\xd3\xd7\x6c\xc7\xaf\x3f\xcb\x77\x37\x6f\xaa\x09\x8d\xb2\xe0\xc3\xd2\x50\xfc\xb0\x3f\x3f\xdb\x95\xd9\xd7\x42\x4c\xfc\xab\x74\x87\xe3\xdb\x0f\xfb\xdb\x97\x49\xdf\x91\xc6\xb3\x94\x1f\xfc\x0d\x9c\xff\x02\xe5\x4f\xc3\x10\xa3\x64\x9e\x55\x69\x56\x45\xc1\x74\x41\x22\x9f\x4e\x72\x42\x72\x3f\x88\x17\x61\x16\x64\x64\x9e\x91\x79\x38\x7d\x91\xf2\xe3\x88\x62\x98\x86\x15\xc9\x93\x8a\x56\x41\x59\x24\x45\x46\xa3\x24\xf5\x53\x46\x8a\x3c\x43\x56\x51\x92\xc6\x65\xf9\x22\xe5\xb3\x79\x40\x93\x79\x10\x05\xd3\x28\x4e\x27\x2c\x9d\xb3\x02\x09\x2b\xf2\x72\x9c\x55\xc1\x62\x31\x1d\x07\x7e\x90\x86\x2c\x7f\x99\xf2\xa3\x3c\x08\x90\xfa\x94\xc4\xcc\x8f\xe3\x28\xc5\x71\x58\xa5\x34\xc4\x71\x3e\x21\xd5\x64\x11\xe4\x49\x3c\x8f\x92\x64\xd1\x53\xfe\xa5\x6c\xb4\xc1\x27\xa4\x5f\xca\x55\x43\x0d\x5b\xff\xb9\xbe\x28\xfc\x8b\x65\x72\xdc\x1d\x7e\xb8\xfe\x65\xf6\x0b\x30\x85\x96\xf3\x55\x6f\xaa\x2d\x15\xa7\xe7\xc7\x67\x2b\xe7\x6f\x6f\x97\xfe\x7f\x0d\x53\x17\x84\xe7\xaa\x27\xfc\xdf\x16\x4f\x94\x96\xa1\x9f\x47\x55\x16\x67\xa5\xbf\xa0\x41\x32\x8d\x7d\x9a\x87\x49\xe8\x4f\xe7\xe3\x04\x17\x34\x4e\xc9\x3c\x0b\xc3\x97\x8b\x27\xf6\x13\x32\xf1\x03\x92\xfb\x2c\xf2\xe7\xe9\x34\xf2\x2b\x32\x19\xc7\x63\xcc\x83\x98\x92\x28\x48\x58\x44\xc7\xf3\xf1\x8b\xc5\x33\x0f\x83\x20\x23\x29\x09\xd3\x74\x82\xc9\x6c\x46\xc2\x2a\xc8\x93\xc4\x1f\xcf\xfc\x3c\x9a\x97\x24\xcc\xaa\x20\x0b\x49\xf5\x07\xfd\x52\x16\x10\x46\xe2\x79\x86\xf3\xb4\x4c\xc2\x34\x20\xb3\x45\x36\x8b\x31\x99\xc6\x79\x96\x8e\x59\x32\xcb\xf3\x6c\x1a\x05\x27\x03\x7b\xd9\xa4\x86\xc2\x95\x91\x8a\xae\x70\xa0\xbb\xff\xdd\x15\x72\x49\xcd\xda\x85\xb8\xb6\x37\x91\xd9\x04\x2a\x5e\xe3\xc0\x6e\x6a\xd6\x23\x38\x35\x9b\xe6\xf4\xf7\xab\xec\xbf\x4a\x6a\xe8\xd0\xad\x2c\x0b\xab\x77\x2a\x45\xc5\x57\xad\x72\x66\xdd\x6f\xc0\xdc\xe8\xd5\x9f\xdf\xa6\x53\xf0\x64\xb7\x31\x63\xb2\x15\x46\xc3\x1d\x1e\xa0\xf7\x62\x40\xfb\x41\xbb\xcf\x1d\x1e\xec\x30\xf6\x1a\x8f\x53\x56\xf6\xed\x7d\x4f\xb0\xb3\x48\x74\x88\x1a\x2f\xdf\x02\x15\x25\x2c\x83\x25\x5c\x75\x07\xba\x2d\x7e\x14\xb6\xba\x07\xb6\x6e\xdf\x48\x6d\x04\xdd\xe0\x08\x88\xbb\x7c\x92\xc1\x2b\x58\x4a\x65\x7a\x25\x56\xc1\xf7\x05\xed\xa2\x11\x64\x24\x0b\xec\xe6\xb6\xdc\x3d\x23\x5d\x4f\x04\xec\x61\xcc\xf4\xa0\x09\x9a\x2e\x44\x57\x0d\x32\x5e\x1d\x60\xbe\x37\xee\xe8\x85\xb7\xcb\x07\xb6\xba\x5e\x81\x51\x61\xaf\xf2\x0a\x6d\x3b\x54\x02\x35\xc0\x2b\x28\x70\xcd\x45\x09\xe7\xe3\x6b\xab\x06\x7b\xe9\xb7\xcb\x11\xec\x86\xfb\xe1\x61\xf8\xb5\x4b\x80\xb5\xba\xd5\x58\xde\xd7\x93\xf5\xba\xa6\x07\x54\x36\x0d\xce\x5c\xc7\x06\x6e\xf5\x35\xdf\xa0\x6c\x9d\x9b\x02\x64\x83\xa2\x7f\x5f\xe8\x9b\x21\xc7\x7e\xae\xc1\x1b\xc0\x71\xb8\x17\x19\xc1\x49\x48\xb4\x03\xdd\x45\x8b\x2d\x7e\xe3\xae\xdb\x9d\xea\x83\x60\x6b\x25\x85\x6c\xb5\x25\x54\x86\x5a\x73\xb1\x1a\x7c\xb1\x02\x5d\x30\xba\xd7\x11\xdd\xb9\xde\x6e\x0a\x54\x96\x92\x2d\x75\xa0\xd2\xa7\x4c\x0a\x6d\x59\xbe\xa7\xe7\x9d\xbd\x94\x16\xae\xdb\x93\x8c\x9a\x2e\x32\xda\x50\x65\xda\x66\x00\x56\xfe\xa6\x13\x1c\x41\xe7\xde\x42\x21\x6a\x68\x1b\x98\x2e\x3f\x00\x3b\xb0\x1a\x75\xe7\x6a\xb7\x81\x6d\xe4\x77\x94\xbb\x47\x15\x6b\x2f\x6e\xd1\xa2\x08\xfa\xe9\x1b\xca\x9d\xb7\xef\xaf\x46\xe0\x0f\xfa\x23\xa7\xb7\x50\xa1\x51\x1c\x5d\x43\x2a\x77\x7d\xb0\x29\x18\xaa\xed\x91\x63\xff\x5d\x76\x0b\x46\xe0\x13\x1b\xa3\x7b\xe6\xd4\x2e\xfb\x9c\x3d\x8e\xd7\xe0\xc8\x9b\x3d\x44\xb0\x46\x4b\x89\xbb\x35\x67\xeb\x7b\x4e\x85\x1e\xe7\x36\x29\xf6\x42\xd2\x9f\x7a\xd2\xc6\xaf\x3f\xae\x4a\xe0\x5d\xe7\xc9\x5a\x6d\xe4\xa6\xdf\xe4\x58\x84\xfd\xfb\x53\x5f\x5e\xe7\x0e\xef\x27\x1b\xca\xc5\xc9\xfd\x2b\x93\xab\xef\x5e\xf1\xfd\xbe\xac\xb6\x77\x85\x0e\x9a\x3f\xec\xd0\x5d\x95\xb8\x42\xd8\x69\x90\x0a\x78\xc3\xfa\xa7\x27\x5a\xd4\x68\x3f\x99\x3b\x28\xbb\x68\xda\x03\xd1\x0a\x7e\xb8\x3c\x1b\xc1\xda\x98\x66\x74\x7a\xea\x7a\x73\xdb\xd0\x8f\xf2\x38\x8a\x8f\x38\x70\x4f\x63\x2b\x6a\x7d\xe1\xcc\x9a\xbb\xa2\x7a\x69\x3f\x6d\x0c\x8f\x7f\x4f\x16\xd7\x7c\xc3\x4d\xb7\xf8\xcc\x7e\x8e\x20\x4a\xfd\x20\xcc\xb2\x47\xf8\x36\xd2\x25\xba\x4b\x93\xf8\xdd\x33\xa3\xa8\xd0\xf4\xbe\xf1\xb7\x3e\x94\x65\xf7\x94\x46\xc1\xdd\x8d\x1c\x71\x74\xae\x80\x51\x7c\xb5\x42\x85\x65\x57\x0d\x06\xf7\xe6\x88\x91\xae\x22\x12\x62\x4b\xe2\xb9\x8d\x15\xd2\x12\xa4\xa8\x0f\xb6\xd2\x8e\x75\x72\x7c\x4f\x3c\x9a\xf4\xbb\xea\x4b\xa4\xe5\x63\xf5\x7e\xdc\x6b\x3f\xb7\x99\x78\x68\x7b\x23\x65\x0d\x1b\xba\xbf\xc7\xa5\x91\xa0\x51\x94\x16\x93\x0f\x96\xc9\xad\x63\x81\x0d\xdd\xdf\xc3\x33\xe8\x63\xfa\x7d\x95\xee\x86\xb5\xa5\xb5\xd3\x7b\xe8\x6a\x87\x5a\x03\x59\xab\x94\x7b\xcb\x7a\x20\xb1\xa6\x1a\x0a\x44\x01\x25\x1a\x64\xc6\x85\xe9\xa8\xc0\xee\x67\x4f\xc5\xa0\xf7\x60\xc6\xb5\x43\x8b\xd3\xa8\xe5\xe6\x09\xda\x34\x94\xf2\xe1\x45\x1c\xcc\xde\x59\x44\x1b\x6e\x2b\x6c\xbf\x94\xb2\x1e\x33\xcb\x28\x73\x61\x35\x95\x23\x30\xaa\x45\x5b\x6b\x54\x1c\xa0\xc4\xa2\x5d\xad\x7a\x36\xb3\x25\xe0\xb8\x63\x25\xc1\x6e\x32\x70\xb3\x5d\xa9\x35\x8d\x92\x95\x4b\xcf\xbd\x88\xe5\x49\x3b\x3a\x82\x8a\xd6\x1a\x07\xff\x0e\x00\x00\xff\xff\x34\x74\x03\xb1\x11\x16\x00\x00")
+var _goCentrifugeBuildConfigsDefault_configYaml = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\xc4\x58\x59\x73\x1b\x37\x12\x7e\xe7\xaf\xe8\x92\x5f\x92\xaa\x1d\x6a\xee\x83\x55\xa9\x2d\xea\xf0\x11\xcb\x0a\x75\x45\xb1\x5e\xd6\x18\xa0\x41\xc2\x1a\x02\x63\x00\xc3\xc3\xbf\x7e\x0b\x98\x21\x23\x59\x96\xb2\x9b\x54\x76\xf5\xa2\x21\x80\x6e\x34\xba\xbf\xef\xc3\xf1\x0a\x4e\x90\x93\xae\xb1\xc0\x70\x85\x8d\x6a\x97\x28\x2d\x58\x34\x56\xa2\x05\x32\x27\x42\x1a\x0b\x5a\xc8\x7b\xac\xb7\x23\x8a\xd2\x6a\xc1\xbb\x39\x9e\xa3\x5d\x2b\x7d\x3f\x01\xdd\x19\x23\x88\x5c\x88\xa6\x19\x79\x67\x42\x22\xd8\x05\x02\x1b\xfc\xca\x7e\xa4\x01\xbb\x20\x16\x8e\xf7\x1e\x60\x49\x84\xb4\xce\xff\x68\x37\x64\x32\x02\x78\x05\x67\x8a\x92\xc6\x87\x20\xe4\x1c\xa8\x92\x56\x13\x6a\x81\x30\xa6\xd1\x18\x34\x20\x11\x19\x58\x05\x35\x82\x41\x0b\x6b\x61\x17\x80\x72\x05\x2b\xa2\x05\xa9\x1b\x34\xe3\x11\xec\xec\x9d\x4b\x00\xc1\x26\x90\x24\x89\xff\x46\xbb\x40\x8d\xdd\x72\x58\xc1\x3b\x36\x81\x32\x29\xfb\xbe\x5a\x29\x6b\xac\x26\xed\x0c\x51\x9b\xde\x36\x80\x83\x43\xd1\xa6\x87\x51\x5c\x8c\xc3\x71\x38\x8e\x0e\x2d\x6d\x0f\x93\x32\x0e\xe3\x43\xd1\x72\x73\x78\xb1\xbc\xbe\xd8\xd4\xeb\xfb\xee\xee\xe3\xc7\x13\xde\x7d\xbd\xae\x37\xa7\xd3\x4b\xbc\x3e\x3f\x3e\x53\x5f\xb7\xdb\x2c\x2b\x57\x17\x72\xfe\xeb\x6a\xf6\xe1\xf3\xd9\xc7\xfb\x83\x3f\x70\x9a\xec\x9c\xfe\xca\xf3\xd3\xf3\x7c\x79\xff\xe5\x16\x3f\xdf\xbe\xbf\x8d\xbf\xcc\xba\x28\xff\xad\x65\x6f\x92\xfb\x9f\x55\x74\x9d\x2c\x17\x64\x31\x3b\xca\xae\x30\x93\x51\xef\x74\x97\xaa\xe9\x2e\x53\xfd\x02\xdc\xf2\x51\x5a\x61\xb7\xaf\x09\xb5\x4a\x6f\x27\x70\x70\xf0\x4d\xcf\x25\xce\x85\xb1\x8f\xba\x88\xa4\x0b\xa5\x2f\xb1\x55\x46\x7c\x63\xd5\x92\xad\x83\xc9\x2f\x75\x23\xe6\xc4\x0a\x25\x7d\x9f\x2f\xde\x07\x22\xe4\x77\xa1\x34\xd4\x18\x7e\xb8\xec\xb1\xf4\xe3\x08\x1e\x62\xa7\x0f\xf5\x15\x9c\x77\x4b\xd4\x82\xc2\xbb\x13\x50\xdc\xe3\xe8\x01\x62\x06\x1f\xfb\x92\x66\xd1\x60\x75\xb4\xab\x1b\x34\xc2\x58\x67\x29\x15\xc3\xa7\x90\x6b\xb5\x5a\x09\xdf\xa1\xbc\xef\x07\x01\xec\x02\xfd\x43\x1c\x24\xd9\x38\x8e\xb3\x71\x1c\x86\xe3\x34\xfe\x16\x0b\x51\x7c\x92\xbc\x57\xea\xf6\x4c\x08\x7a\xf1\xeb\xfa\x7a\x71\x7d\xf4\x31\xdf\xbc\xa7\x33\x75\xc6\xf3\xcb\x8b\x8f\x3f\xbf\x6e\xd7\x3c\xd2\x45\xb6\x3e\xdb\xc4\x77\x97\x49\x7b\xcc\xa2\x83\xef\xb9\x2f\xf3\x71\x1c\x85\xcf\xb9\xbf\xb8\xfb\x30\x2d\xdf\xcc\xde\xea\xd5\xe9\xdd\x51\xb5\x66\xf7\xea\x86\x4e\xa7\xcb\xe3\xbb\xb7\x6d\x85\xdb\xed\x5d\x7a\x75\x5a\xce\x5f\xeb\x64\x71\x7d\xfe\xdb\xc1\x90\xa3\xd3\x01\xf7\xfb\x4a\xbc\x3b\x81\x00\x86\x6a\x3c\xc7\x8c\x74\x30\x3e\x23\x2e\x3d\xc0\xb0\x6d\xd4\x16\x19\x5c\x2d\x89\xb6\x70\x3c\x00\xce\x00\x57\xda\x27\x74\x2e\x56\x28\x1f\xa5\xf2\x29\x28\xe1\x59\x54\x86\x9b\x8c\xc4\x69\x95\x17\x9c\xb3\xb4\x24\x31\x89\x68\x99\x30\x56\x66\x49\xc6\xea\x92\xd5\xac\x62\x24\xca\x73\x9a\xe4\x2f\xe0\x37\xdc\x54\x79\x1e\xd2\x30\xa9\x58\x12\x45\x69\x96\x10\x1e\xb2\xac\xa4\x59\x9e\xe7\x45\x9c\xb0\x8a\xc6\x9c\x14\x2c\x47\xfa\x02\xd2\xc3\x4d\x86\x29\x61\x69\x11\xe5\x65\x9e\xc4\x51\x5e\x57\x09\xe5\x49\x8c\x51\xcc\xf2\x8a\x53\x1e\x15\x45\x9c\xd5\x59\xc2\x5e\xe2\x44\xb8\xc9\x6b\x96\xf0\x2c\x09\x8b\x88\x55\xbc\xac\x38\xa5\x59\x15\x46\x59\x95\xd7\x45\x4a\x93\x9a\x45\xb4\xa8\xa2\xb0\x48\x07\xf6\xbc\x57\x2b\xd2\xa7\xef\x01\xd6\x6b\xd4\x92\x34\x0b\x14\xf3\x85\x1d\xb0\xf8\xea\xd5\xab\xa1\x30\xbd\xc5\xeb\xe9\xc5\xf0\x3b\x80\x5b\x27\x87\x42\xf2\x4e\x13\xd8\xaa\x0e\xe6\x4e\xc7\x25\xa0\xd6\x4a\x3b\x94\x5d\x2f\x84\x01\x8d\x5f\x3a\x37\x8b\x30\x20\x95\x05\xd3\xb5\xad\xd2\x16\x19\xd4\x48\x49\x67\xd0\x59\x6a\x4f\x22\x37\x44\x77\x52\x3a\x2d\xf6\x4a\x6b\x2c\xb1\x8e\x49\x9d\x6b\x1a\xc3\x65\x27\xfb\xf6\x20\x18\xda\x7e\x22\x9a\x2e\xc4\x0a\xc7\x07\xff\x18\x82\x02\x58\x3b\x22\x5a\x05\x4c\xfd\xd3\x5b\x10\x68\xbc\xca\xb7\x44\x0b\xbb\xed\x27\xf2\x5e\xee\xfd\x7a\x70\x3e\xe9\x7f\x7e\x1a\x06\x04\x01\x5d\x10\x21\x7f\xea\xbb\x83\xc0\x45\xfb\x53\x12\x26\x61\x0a\x41\xb0\x26\xba\x1d\xfe\x05\x35\xd1\x5a\xa0\x86\x2c\x2f\xc3\x30\x0c\x21\x08\xa4\x0a\x88\xa4\x02\xa5\x0d\xea\x46\xd1\x7b\xd3\xb7\x19\xd4\x2b\x0c\x1a\x97\x54\x08\x82\x25\xd9\x04\xad\xe3\x3a\xc4\x99\x33\x32\x92\xb4\x66\xa1\xec\xd0\xe8\xdb\x96\x42\x3e\xfa\xe9\x62\x26\xd4\x8a\x15\x42\x10\x38\x8c\xbb\x14\x29\xce\x9f\x66\x02\x82\x80\xd5\x01\x55\xcb\xd6\x8d\x57\x12\x8c\x61\x6e\x49\x84\x2e\x30\x30\xe2\x2b\x42\x1a\x56\x39\x04\xc1\x67\xa3\xa4\x6e\x69\xb0\x50\xc6\x1a\x20\x4d\xf3\xa0\x4d\x48\x8b\x9a\x13\x8a\xae\xfd\xd3\xe3\x72\x3f\x4d\xe6\xf7\x2a\x7f\xe4\x96\x8f\xcc\x51\x52\x62\x1f\x88\x55\x70\x8b\xf5\x95\x6b\xb7\x06\x7c\x4e\x34\x70\xad\x96\xd0\x49\xab\x3b\xe3\x20\xa1\xb4\x98\x0b\x39\x81\xf1\xf8\xe0\xd9\x7a\x3a\xee\x3f\xa9\xe5\xa7\x20\xe8\xa4\x21\x1c\x03\xdc\xb4\xca\xe0\x27\xe0\x0d\x99\x7f\x03\xe0\xff\x4e\xf0\xe3\xbf\x28\xf8\x8f\xb8\xf4\x1f\x4b\x7e\x14\xa6\xe3\x28\x4b\xc7\x51\x39\xce\x9e\x6c\xff\x3b\x4d\x9e\x99\x5c\x10\xbc\xe9\x5e\xdf\x9d\x77\xd1\x9b\xcd\xca\x6c\x8f\xae\xaf\xf4\xb5\xa9\x56\xf6\x28\xaf\xed\x87\xa9\x7c\xfb\x5a\x9d\x7d\xae\xef\xbf\x1e\x93\x83\xef\xb8\xcf\xc6\x51\x99\x8d\xe3\xa4\x78\x76\x82\xe3\x37\x74\x2d\xae\x3f\xab\xf7\xb7\x6f\xf9\x11\x49\xcb\xf8\x66\x66\x09\xde\x6c\xce\xcf\xd6\xac\xfc\x5a\xcb\xa3\xe8\xaa\x58\xe3\xf4\xee\x66\x73\xf7\xb2\xe8\x7b\xd1\x78\x56\xf2\xe3\xbf\x41\xf3\x5f\x90\xfc\x98\xb3\x8c\xf3\xb2\x62\xc8\xeb\x38\x23\x21\x29\xb0\x8e\x39\xc6\x2c\x42\xa4\x61\x1c\x95\x45\x9d\xe5\x18\x25\x2f\x4a\x7e\x96\x12\x4c\x8a\x84\x87\x55\xce\x09\x8f\x59\x9d\xd7\x25\x49\xf3\x22\x2a\x68\x58\x57\x25\x52\x4e\xc2\x22\x63\xec\x45\xc9\xc7\x9a\xb3\x12\x13\x92\x87\x65\x98\xb1\x34\x8c\x29\xab\xc3\x3c\xa5\x75\x56\xd7\x45\xc5\xe2\x34\x2e\xa3\x88\xc5\xf4\x65\xc9\xe7\x55\x92\x54\xac\x4a\xa2\xa8\x26\x71\x59\xd2\x22\x2f\xcb\xb0\x8a\x79\x5d\xc6\x09\x2f\x63\x96\x44\x59\x9e\x72\x5e\x0f\x92\x7f\xa9\x5a\x63\xf1\x89\xe8\x33\x35\x6f\x89\xa5\x8b\x3f\x77\x2e\x4a\xfe\x22\x4d\x76\xb3\xc3\x0f\xd7\xbf\x9c\xfc\x02\x54\xa3\xd3\x7c\x3d\x84\xea\xa8\xe2\xfd\xfc\xf8\x2c\x73\xfe\xf6\xe3\xd2\xff\xef\xc0\xd4\x27\xe1\x39\xf6\x24\xff\x5b\xf2\x54\xbc\x8a\x79\x18\x92\x84\x65\x05\x2b\x78\x12\xe5\x18\x53\x12\x97\x55\x95\x33\x9e\xc5\x55\x54\x66\x21\x96\x84\xbe\x4c\x9e\x2c\xca\xc3\xa3\x28\x0e\xab\x88\xa6\xd1\x69\x71\x9c\x46\x3c\x3c\x9a\x66\x53\xac\x1c\x21\xd3\x38\xa7\x29\x99\x9e\x4e\x5f\x24\x0f\xcf\x59\x4e\x12\xe4\x55\x8c\x55\x15\x21\xe6\x75\xc1\x49\x54\xd2\xb8\x08\x63\x92\x25\x71\x1e\x96\x31\x21\xac\x7a\x99\x3c\x69\x52\x84\x59\xc6\x6a\xc6\xd3\x2c\x2e\x13\x86\x29\xc9\xb1\xc2\xb0\xaa\xab\x10\xc3\x34\x21\x69\x49\x63\x56\x47\x07\x23\x77\xd9\x24\x96\xc0\x95\x55\x9a\xcc\x71\x64\xfa\xff\xfd\x15\x72\x46\xec\xc2\xa7\xb8\x71\x37\x91\x93\x23\xe0\xa2\xc1\x91\x9b\xd4\x2e\x26\x70\x68\x97\xed\xe1\xef\x57\xd9\x7f\x31\x62\xc9\xd8\x8f\x64\xb5\xf3\x7b\xac\x24\x17\xf3\x4e\xfb\xb0\xf6\x13\x50\xdf\x7a\xf5\xe7\xa7\xe9\x1d\x3c\x99\x6d\x4a\xa9\xea\xa4\x35\x70\x8f\x5b\x18\x56\x31\x22\x43\xa3\x9b\xe7\x1e\xb7\xae\x19\x07\x8f\xbb\x2e\x67\xfb\x6e\x7f\x26\x58\x3b\x24\x7a\x44\x4d\x67\xef\x80\x48\x06\xb3\x78\x06\x57\xfd\x86\xee\xc8\x8f\xd2\xb1\x7b\xe4\x78\xfb\x56\x19\x2b\xc9\x12\x27\x10\xfa\xcb\x67\x38\x7a\x05\x33\xa5\xed\xe0\xc4\x39\xf8\xbe\xa1\x1b\x34\x81\x32\x2c\x63\x37\xb9\xa3\x7b\x60\x95\x3f\x13\x01\x7d\x98\x33\x33\x6a\xe3\xb6\x4f\xd1\x55\x8b\x54\xf0\x2d\x9c\x6e\xac\xdf\x7a\xe1\xdd\xec\x41\xac\xfe\xac\x40\x89\x74\x57\x79\x8d\xee\x38\xc4\x80\x58\x10\x1c\x6a\x5c\x08\xc9\xe0\x7c\x7a\xed\xdc\xe0\x60\xfd\x6e\x36\x81\xf5\x78\x33\xde\x8e\xbf\xf6\x05\x70\x51\x77\x06\xd9\x9e\x4f\x6e\xd5\x0d\xd9\xa2\x76\x65\xf0\xe1\x7a\x35\xf0\xa3\xaf\xc5\x12\x55\xe7\x97\x29\x41\xb5\x28\x87\xf7\x85\xe1\x30\xe4\xd5\xcf\x1f\xf0\x46\xb0\x6b\x1e\x4c\x26\x70\x90\x84\xc6\x83\xee\xa2\xc3\x0e\xbf\x59\xae\x9f\x9d\x98\xad\xa4\x0b\xad\xa4\xea\x8c\x13\x54\x8a\xc6\x08\x39\x1f\x7d\x71\x06\x7d\x32\xfa\xd7\x11\xd3\x2f\xbd\x5b\xd6\xa8\x9d\x24\x3b\xe9\x40\x6d\x0e\xa9\x92\xc6\xa9\xfc\x20\xcf\x6b\x77\x29\xad\xfd\x69\x4f\x51\x62\xfb\xcc\x18\x4b\xb4\xed\xda\x11\x38\xfb\xdb\xde\x70\x02\xfd\xf2\x5e\x6b\x44\x03\x5d\x0b\xc7\xb3\x1b\xa0\x5b\xda\xa0\xe9\x97\xda\x4f\xe0\x0e\xf2\x6b\x22\xfc\xa3\x8a\x8b\x17\x57\xe8\x50\x04\x43\xf7\x2d\x11\x7e\xb5\x1f\xae\x26\x10\x8d\x86\x2d\x67\x88\x50\xa3\xd5\x02\xfd\x81\x54\xad\x87\x64\x13\xb0\xc4\xb8\x2d\xc7\xfd\xbb\xec\x07\x4c\x20\x0a\x5d\x8e\xf6\xca\x69\x7c\xf5\x05\x7d\x9c\xaf\xd1\x4e\x37\x07\x88\x60\x83\x4e\x12\xd7\x0b\x41\x17\x7b\x4d\x85\x01\xe7\xae\x28\xee\x42\x32\xec\x7a\xca\xe5\x6f\xd8\xae\x18\x88\xfe\xe4\x49\x3b\x63\xd5\x72\x98\x64\x47\xc2\xe1\xfd\x69\xa0\xd7\xb9\xc7\xfb\xc1\x92\x08\x79\xb0\x7f\x65\xf2\xfc\x1e\x1c\xef\xe7\xa5\x8d\xbb\x2b\xf4\xd0\xfc\x61\x8d\xfe\xaa\x24\x34\xc2\xda\x80\xd2\x20\x5a\x3a\x3c\x3d\x91\xba\x41\xf7\x49\xfd\x46\xd9\x67\xd3\x6d\x88\xce\xf0\xe6\xf2\x6c\x02\x0b\x6b\xdb\xc9\xe1\xa1\x3f\x9b\xbb\x03\xfd\xa4\xca\xd2\x6c\x87\x03\xff\x34\x36\x27\x6e\x2d\x82\xba\x70\xe7\xc4\xcc\xdc\xa7\xcb\xe1\xee\xef\xc9\xe0\x46\x2c\x85\xed\x07\x9f\xb9\xcf\x09\xa4\x45\x14\x27\x65\xf9\x08\xdf\x56\xf9\x42\xf7\x65\x92\xbf\xaf\xcc\x6a\x22\x0d\xd9\x1f\xfc\xdd\x1a\x18\xeb\x9f\xd2\x08\xf8\xbb\x91\x17\x8e\x7e\x29\x60\xb5\x98\xcf\x51\x23\xeb\xd9\x60\x71\x63\x77\x18\xe9\x19\x91\x87\x8e\x12\xcf\x4d\xac\x91\x30\x50\xb2\xd9\x3a\xa6\xed\x78\xb2\x7b\x4f\xdc\x85\xf4\xbb\xeb\x4b\x24\xec\xb1\xfb\x28\x1b\xbc\x9f\xbb\x4a\x3c\x8c\xbd\x55\xaa\x81\x25\xd9\xec\x71\x69\x15\x18\x94\xcc\x61\xf2\xc1\x30\xb5\xf2\x2a\xb0\x24\x9b\x3d\x3c\xe3\x21\xa7\xdf\x77\xe9\x6f\x58\x2b\xd2\x78\xbf\xdb\x9e\x3b\xc4\x05\x48\x3b\xad\xfd\x5b\xd6\x03\x8b\x05\x31\x50\x23\x4a\x60\x68\x91\x5a\x9f\xa6\x9d\x03\x37\x9f\xdb\x15\xe3\x61\x05\x27\xc2\x78\xb4\x78\x8f\x46\x2d\x9f\xa0\xcd\x00\x53\x0f\x2f\xe2\x60\x37\x3e\x22\xd2\x0a\xc7\xb0\xcd\x4c\xa9\x66\x4a\x9d\xa2\x9c\x4a\xe7\x89\x4d\xc0\xea\x0e\x1d\xd7\x88\xdc\x02\xc3\xba\x9b\xcf\x07\x35\x73\x14\xf0\xda\x31\x57\xe0\x26\x19\xf9\xde\x9e\x6a\x6d\xab\x15\xf7\xe5\xd9\x9b\x38\x9d\x74\xad\x13\xe0\xa4\x31\x38\x1a\xf5\xbb\xfb\xf0\x74\xda\x6a\xa4\x6a\xe9\x91\xe6\x27\xfc\x77\x00\x00\x00\xff\xff\xf3\x86\xba\x5a\x2f\x16\x00\x00")
 
 func goCentrifugeBuildConfigsDefault_configYamlBytes() ([]byte, error) {
 	return bindataRead(
@@ -84,12 +84,12 @@ func goCentrifugeBuildConfigsDefault_configYaml() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "go-centrifuge/build/configs/default_config.yaml", size: 5649, mode: os.FileMode(420), modTime: time.Unix(1551103801, 0)}
+	info := bindataFileInfo{name: "go-centrifuge/build/configs/default_config.yaml", size: 5679, mode: os.FileMode(420), modTime: time.Unix(1551789277, 0)}
 	a := &asset{bytes: bytes, info: info}
 	return a, nil
 }
 
-var _goCentrifugeBuildConfigsTesting_configYaml = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\x94\x53\x3b\x8f\xdc\x36\x10\xee\xf5\x2b\x08\xa6\xb8\x66\x1f\x7c\xbf\xba\x94\x81\x91\x34\x09\xe0\x7a\x48\x0e\xd7\xc4\xde\x6a\x15\x92\xba\xf3\xc2\xf0\x7f\x0f\xb4\xde\x4b\xae\xcb\xb9\x9b\x19\x7d\x8f\x19\xe9\x53\xc2\x79\xb4\x5a\xd6\x13\xfe\x81\xe3\xf5\xda\xce\x81\x0c\xec\xa3\xce\xa7\x09\xc7\x17\x6c\xb8\x5e\xc2\x44\x08\xa4\x74\x5d\xe7\xd1\xb7\x9a\x90\x0b\xd4\x39\x90\x7b\x49\xc8\x19\x6f\x81\x3c\x7d\xa3\x90\x73\xc3\xde\x69\xa0\xce\x47\x06\xce\x68\x27\x93\x52\x4a\x41\x2a\xd9\xf2\xa8\x8c\x44\x96\x65\xd2\x1a\x90\x2b\x2e\x40\xd3\x1d\x4d\xed\xb6\x8c\x2b\x0d\xdf\x68\xaa\xcb\x17\x6c\x34\x50\xc0\xbe\xe7\xc2\xed\xd3\x68\x1b\xe0\x3e\x1e\xf8\x75\xd0\x40\x93\xb5\xbe\x38\x69\x7d\xb6\x96\x65\x2f\x52\x49\x3c\xe7\xac\xc0\x15\xc9\xb3\x06\x06\x39\xb9\x22\x80\x45\x01\x5c\x31\x2e\x2d\xcb\xd2\x48\x56\xa4\x4b\x2c\x39\xf8\x57\x6f\x81\x06\x97\xbe\xd9\xd6\x17\x1a\xa8\x34\x89\x1b\x87\x56\xc6\xe2\x1d\x2b\x68\x75\x64\x56\xd8\xe2\x3c\x03\xcb\x21\xd3\xef\x3b\x7a\xce\x85\x06\xda\xef\x0b\xd3\x7b\xfb\x9f\x48\x3e\x3f\xe3\x4c\x83\x14\x3b\x3a\xd3\x20\x8c\xe0\x4a\xed\xe8\x42\x03\xdf\xd1\x46\x83\xdb\xd1\x0e\xcf\xdb\x01\x19\x79\x44\x6e\x50\x26\xef\xb8\x57\x2a\x73\x4c\x20\xa2\x8b\xc2\xa2\x42\x83\x2c\xea\x58\xa2\x92\x11\x99\xb4\x06\x74\x76\xce\xf9\x02\xc6\x7a\x10\x8e\x0b\xb1\x2d\x72\x81\xb4\xbd\x8a\xc4\x85\x8b\x8e\x6b\xad\x75\x04\x8e\x90\x6d\x02\xf4\xcc\x30\x74\x4e\x09\x28\x09\x9c\xd4\x26\x33\xa3\xb4\x8e\xd9\x83\xb6\x5a\x44\x30\x25\x25\xe6\x05\x96\x4d\xa9\x66\x1a\xa8\xd2\xc8\x0c\x03\xb3\xcf\x02\x70\xaf\x64\x74\x7b\x2f\x44\xd9\x2b\xe5\x84\x57\xde\x67\x69\x33\xdd\xd1\x17\x6c\xbd\x5e\xb7\x23\xbf\x3f\x3d\x3e\xfc\x02\xbd\xbf\x5e\x5b\x0e\xe4\xe9\x6d\xf4\xc8\x40\x20\x1f\x8d\xc0\x34\xd5\x8c\xf3\xa8\xe3\xf6\x5b\x0e\x84\xb2\xaf\x1f\xce\xce\x34\xfd\x42\x7e\x7d\xa4\x72\xcb\x20\xe9\xe3\xda\xe0\x84\xd3\xfb\xa8\x9e\xf1\xb6\x8d\x31\x90\xe3\xb8\x2c\xc7\xb7\x47\xd3\xf4\xf7\x8a\x2b\x6e\x88\x79\xbd\x7c\xbe\xb6\x33\xb6\x1e\x88\x98\x08\x79\xbd\x37\x9f\xa1\x8e\xbf\xea\x05\x7f\xff\x33\x10\x3e\x4d\x9b\xcc\x06\x5e\xc4\xf2\xe3\x07\x58\xd6\xf8\x5c\xd3\xa7\x2d\xf9\x87\xc3\xf1\x70\x38\xc6\xb5\x3e\xe7\x63\xc3\x7e\x5d\x5b\xc2\x7e\x5c\xc4\xf2\x09\x6f\x87\x65\x8d\x87\x05\x2f\x3f\x38\xad\xbe\xc0\xc0\xff\x27\x9d\x37\xe2\x9d\xd4\xeb\x69\xae\xf3\xe9\x83\x9e\x0f\xf4\xcf\xfb\xbe\x23\xbe\x79\xff\x13\x00\x00\xff\xff\x59\xcf\xd7\x6b\x17\x04\x00\x00")
+var _goCentrifugeBuildConfigsTesting_configYaml = []byte("\x1f\x8b\x08\x00\x00\x00\x00\x00\x00\xff\x94\x53\xb9\x8e\xe4\x36\x10\xcd\xf5\x15\x04\x1d\x4c\xd2\x07\x4f\xf1\xc8\x1c\x1a\x0b\x3b\xb1\x81\x8d\x8b\x64\xb1\x87\xe8\xd1\x61\x92\x9a\xd9\xc6\x62\xff\xdd\x50\x6f\x8f\xbd\x99\x67\xb3\xaa\xd2\x3b\xaa\xa4\xa7\x88\x73\xaf\x25\x6f\x17\xfc\x03\xfb\xdb\x52\xaf\x9e\x74\x6c\xbd\xcc\x97\x01\xfb\x33\x56\xdc\x26\x3f\x10\x02\x31\x2e\xdb\xdc\xdb\x5e\x13\x32\x41\x99\x3d\xb9\x97\x84\x5c\xf1\xe6\xc9\xd3\x57\x0a\x29\x55\x6c\x8d\x7a\x6a\x5d\x60\x60\x47\x6d\x65\x54\x4a\x29\x88\x39\x19\x1e\xd4\x28\x91\x25\x19\xb5\x06\xe4\x8a\x0b\xd0\xf4\x40\x63\xbd\xad\x7d\xa1\xfe\x2b\x8d\x65\x7d\xc6\x4a\x3d\x05\x6c\x47\x2e\xec\x31\xf6\xba\x03\xee\xe3\x8e\x5f\x3a\xf5\x34\x1a\xe3\xb2\x95\xc6\x25\x63\x58\x72\x22\xe6\xc8\x53\x4a\x0a\x6c\x96\x3c\x69\x60\x90\xa2\xcd\x02\x58\x10\xc0\x15\xe3\xd2\xb0\x24\x47\xc9\xb2\xb4\x91\x45\x0b\xff\xea\xad\x50\x61\x6a\xbb\x6d\x79\xa5\x9e\xca\x31\xf2\xd1\xa2\x91\x21\x3b\xcb\x32\x1a\x1d\x98\x11\x26\x5b\xc7\xc0\x70\x48\xf4\xdb\x81\x5e\x53\xa6\x9e\xb6\xfb\xc2\xf4\xde\xfe\x27\x92\xae\x2f\x38\x53\x2f\xc5\x81\xce\xd4\x8b\x51\x70\xa5\x0e\x74\xa5\x9e\x1f\x68\xa5\xde\x1e\x68\x83\x97\xfd\x80\x84\x3c\x20\x1f\x51\x46\x67\xb9\x53\x2a\x71\x8c\x20\x82\x0d\xc2\xa0\xc2\x11\x59\xd0\x21\x07\x25\x03\x32\x69\x46\xd0\xc9\x5a\xeb\x32\x8c\xc6\x81\xb0\x5c\x88\x7d\x91\x09\xe2\xfe\x2a\x22\x17\x36\x58\xae\xb5\xd6\x01\x38\x42\x32\x11\xd0\xb1\x91\xa1\xb5\x4a\x40\x8e\x60\xa5\x1e\x13\x1b\x95\xd6\x21\x39\xd0\x46\x8b\x00\x63\x8e\x91\x39\x81\x79\x57\x2a\x89\x7a\xaa\x34\xb2\x91\xc1\x78\x4c\x02\xf0\xa8\x64\xb0\x47\x27\x44\x3e\x2a\x65\x85\x53\xce\x25\x69\x12\x3d\xd0\x57\xac\xad\x2c\xfb\x91\xdf\x9e\x1e\x1f\x7e\x85\xd6\xde\x96\x9a\x3c\x79\x7a\x1f\x3d\x32\xe0\xc9\x47\x23\x30\x0c\x25\xe1\xdc\x4b\xbf\xfd\x96\x3c\xa1\xec\xcb\x87\xb3\x33\x0c\xbf\x90\x5f\x1f\xa9\xdc\x33\x48\x5a\x5f\x2a\x5c\x70\xf8\x31\xaa\x57\xbc\xed\x63\xf4\xe4\xdc\xa7\xf5\xfc\xfe\x68\x18\xfe\xde\x70\xc3\x1d\x31\x6f\xd3\xe7\xa5\x5e\xb1\x36\x4f\xc4\x40\xc8\xdb\xbd\xf9\x0c\xa5\xff\x55\x26\xfc\xfd\x4f\x4f\xf8\x30\xec\x32\x3b\x78\x15\xeb\xf7\x1f\x60\xdd\xc2\x4b\x89\x9f\xf6\xe4\x9f\x4e\xe7\xd3\xe9\x1c\xb6\xf2\x92\xce\x15\xdb\xb2\xd5\x88\xed\xbc\x8a\xf5\x13\xde\x4e\xeb\x16\x4e\x2b\x4e\xdf\x39\xb5\xbc\x42\xc7\xff\x27\x5d\x77\xe2\x9d\xd4\xca\x65\x2e\xf3\xe5\x83\x9e\x0f\xf4\xcf\xfb\xfe\x40\x7c\xf7\x1e\x60\x8e\xcf\x4b\x7d\x98\xaf\x15\xe3\x32\x4d\xa5\x7b\xd2\xeb\x86\xff\x04\x00\x00\xff\xff\x1f\xaf\xbe\x5d\x34\x04\x00\x00")
 
 func goCentrifugeBuildConfigsTesting_configYamlBytes() ([]byte, error) {
 	return bindataRead(
@@ -104,7 +104,7 @@ func goCentrifugeBuildConfigsTesting_configYaml() (*asset, error) {
 		return nil, err
 	}
 
-	info := bindataFileInfo{name: "go-centrifuge/build/configs/testing_config.yaml", size: 1047, mode: os.FileMode(420), modTime: time.Unix(1551360206, 0)}
+	info := bindataFileInfo{name: "go-centrifuge/build/configs/testing_config.yaml", size: 1076, mode: os.FileMode(420), modTime: time.Unix(1551798935, 0)}
 	a := &asset{bytes: bytes, info: info}
 	return a, nil
 }
@@ -204,7 +204,6 @@ type bintree struct {
 	Func     func() (*asset, error)
 	Children map[string]*bintree
 }
-
 var _bintree = &bintree{nil, map[string]*bintree{
 	"go-centrifuge": &bintree{nil, map[string]*bintree{
 		"build": &bintree{nil, map[string]*bintree{
@@ -262,3 +261,4 @@ func _filePath(dir, name string) string {
 	cannonicalName := strings.Replace(name, "\\", "/", -1)
 	return filepath.Join(append([]string{dir}, strings.Split(cannonicalName, "/")...)...)
 }
+
