+++ utils/tools.go
@@ -2,7 +2,6 @@ package utils
 
 import (
 	"crypto/rand"
-	"crypto/sha256"
 	"encoding/binary"
 	"math/big"
 
@@ -213,16 +212,4 @@ func ConvertProofForEthereum(sortedHashes [][]byte) ([][32]byte, error) {
 	}
 
 	return property, nil
-}
-
-// GenerateHashPair generates a preimage and hash pair
-func GenerateHashPair(preimageSize int) (preimage, hash []byte, err error) {
-	preimage = RandomSlice(preimageSize)
-	h := sha256.New()
-	_, err = h.Write(preimage)
-	if err != nil {
-		return []byte{}, []byte{}, err
-	}
-	hash = h.Sum(hash)
-	return preimage, hash, nil
 }
